requirements.txt
aiosqlite==0.20.0
annotated-types==0.6.0
anyio==4.3.0
argcomplete==2.0.0
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.1.2
black==24.2.0
certifi==2024.2.2
cffi==1.16.0
click==8.1.3
cryptography==42.0.5
databases==0.9.0
distro==1.9.0
ecdsa==0.18.0
exceptiongroup==1.2.0
fastapi==0.110.0
flake8==7.0.0
gitdb==4.0.11
GitPython==3.1.41
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
Mako==1.3.2
MarkupSafe==2.1.5
mccabe==0.7.0
mypy-extensions==1.0.0
openai==1.13.3
packaging==23.2
passlib==1.7.4
pathspec==0.12.1
pipx==1.0.0
platformdirs==4.2.0
psycopg2-binary==2.9.9
psycopg2-pool==1.2
pyasn1==0.5.1
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
pyflakes==3.2.0
pyparsing==3.0.8
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
rsa==4.9
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.28
starlette==0.36.3
tomli==2.0.1
tqdm==4.66.2
typing_extensions==4.10.0
userpath==1.8.0
uvicorn==0.27.1
